asteroids now decay when harvested by a forager

+++ Assets/Scripts/Asteroid.cs
@@ -6,13 +6,13 @@ using Random = UnityEngine.Random;
 
 public class Asteroid : MonoBehaviour
 {
-    public int resource = 50;
+    public float resource = 50;
 
     private void Start()
     {
         var position = transform.position;
         var x = position.x / 1000;
         var z = position.z / 1000;
-        resource = (int)(Mathf.PerlinNoise(x, z) * 100);
+        resource = (Mathf.PerlinNoise(x, z) * 100);
     }
 }
